trash_utilities::parallel::advanced

Function distribute_work

pub fn distribute_work<T, F, R>(data: &[T], worker: F) -> Vec<R>
where
    T: Send + Clone,
    F: Fn(Vec<T>) -> R + Send + Sync,
    R: Send,

Distribute work evenly across available threads.

This function splits a workload into chunks and processes them in parallel. Useful for load balancing when you have a fixed amount of work.

Type Parameters

T: The input data type.

F: The worker function type.

R: The result type.

Parameters

data: The data to process.

worker: Function that processes a chunk of data.

Returns

Vector of results from each chunk.

Examples

use trash_analyzer::parallel::distribute_work;

let data = vec![1, 2, 3, 4, 5, 6, 7, 8, 9, 10];

let results = distribute_work(&data, |chunk| {
    chunk.iter().sum::<i32>()
});
// Results will be sums of chunks