trash_utilities::parallel::data

Function parallel_group_by

pub fn parallel_group_by<T, K, F>(data: Vec<T>, key_fn: F) -> HashMap<K, Vec<T>>
where
    T: Send + Sync,
    K: Send + Sync + Eq + Hash,
    F: Fn(&T) -> K + Send + Sync,

Group elements by a key function in parallel.

This function groups elements of a vector by keys computed from each element, similar to itertools::group_by but with parallel processing.

Type Parameters

T: The element type, must be Send + Sync.

K: The key type, must be Send + Sync + Eq + std::hash::Hash.

F: The key function type, must be Fn(&T) -> K + Send + Sync.

Parameters

data: The input vector to group.

key_fn: Function that computes the key for each element.

Returns

A HashMap where keys are the computed keys and values are vectors of elements.

Examples

use trash_analyzer::parallel::parallel_group_by;
use std::collections::HashMap;

let data = vec![1, 2, 3, 4, 5, 6, 7, 8, 9, 10];
let groups = parallel_group_by(data, |&x| x % 3);

assert_eq!(groups.get(&0), Some(&vec![3, 6, 9]));
assert_eq!(groups.get(&1), Some(&vec![1, 4, 7, 10]));
assert_eq!(groups.get(&2), Some(&vec![2, 5, 8]));