trash_utilities::serde::utility

Function hash_json_ahash

pub fn hash_json_ahash<T: Serialize>(value: &T) -> Result<u64, Error>

Compute a fast non-cryptographic hash of serialized JSON.

This function uses ahash for fast hashing, suitable for hash tables and checksums where cryptographic security is not required.

Type Parameters

T: The type to serialize and hash, must implement Serialize.

Parameters

value: The value to hash.

Returns

Ok(u64) containing the hash value.

Err(serde_json::Error) if serialization fails.

Errors

Returns serde_json::Error if the value cannot be serialized to JSON.

Examples

use trash_analyzer::serde::hash_json_ahash;
use serde::Serialize;

#[derive(Serialize)]
struct Data { id: u32, name: String }

let data = Data { id: 123, name: "test".to_string() };
let hash = hash_json_ahash(&data).unwrap();
println!("Fast hash: {}", hash);