trash_utilities::sys::env

Function read_env_var_or

pub fn read_env_var_or(key: &str, default: &str) -> String

Reads an environment variable with a default value.

Parameters

key: The name of the environment variable.

default: The default value to return if the variable is not set.

Returns

The value of the environment variable or the default.

Examples

use trash_analyzer::sys::env::read_env_var_or;

let port = read_env_var_or("PORT", "8080");
println!("Port: {}", port);